refactor: unify websocket provider
